Refactored ActionMode handling to be extensible. Each ActionMode should now have its own handler, with the respective custom implementation.
Refactored effect handing to be extensible. Each effect can now have its own handler, with the respective custom implementation.

TOOD: ActionMode Enum should be deprecated and refactored to allow extensibility by addons.
