+++ src/main/java/com/dragonminez/server/DMZServer.java
@@ -3,16 +3,18 @@ package com.dragonminez.server;
 import com.dragonminez.Env;
 import com.dragonminez.LogUtil;
 import com.dragonminez.server.commands.*;
-import com.dragonminez.server.world.feature.OverworldFeatures;
+import com.dragonminez.server.events.players.TickHandler;
 import com.mojang.brigadier.CommandDispatcher;
 import net.minecraft.commands.CommandSourceStack;
-import net.minecraftforge.api.distmarker.Dist;
-import net.minecraftforge.api.distmarker.OnlyIn;
 
 public class DMZServer {
 
     public static void init() {
         LogUtil.info(Env.SERVER, "Initializing DragonMineZ Server...");
+
+		// FIXME: Should be moved to one of the mod's initialization events, like Init.
+		TickHandler.registerActionModeHandlers();
+		TickHandler.registerStatusEffectHandlers();
     }
 
 	public static void registerCommands(CommandDispatcher<CommandSourceStack> dispatcher) {
