+++ src/main/java/com/dragonminez/server/events/players/TickHandler.java
@@ -3,34 +3,25 @@ package com.dragonminez.server.events.players;
 import com.dragonminez.Reference;
 import com.dragonminez.common.config.ConfigManager;
 import com.dragonminez.common.config.FormConfig;
-import com.dragonminez.common.config.GeneralServerConfig;
 import com.dragonminez.common.config.RaceStatsConfig;
 import com.dragonminez.common.events.DMZEvent;
 import com.dragonminez.common.init.MainEffects;
 import com.dragonminez.common.init.MainItems;
-import com.dragonminez.common.init.MainSounds;
 import com.dragonminez.common.network.NetworkHandler;
 import com.dragonminez.common.network.S2C.StatsSyncS2C;
 import com.dragonminez.common.stats.*;
-import com.dragonminez.common.util.TransformationsHelper;
-import com.dragonminez.server.util.FusionLogic;
+import com.dragonminez.server.events.players.actionmode.FormModeHandler;
+import com.dragonminez.server.events.players.actionmode.FusionModeHandler;
+import com.dragonminez.server.events.players.actionmode.KaiokenModeHandler;
+import com.dragonminez.server.events.players.actionmode.RacialModeHandler;
+import com.dragonminez.server.events.players.statuseffect.*;
 import com.dragonminez.server.util.GravityLogic;
-import com.dragonminez.server.util.RacialSkillLogic;
 import com.dragonminez.server.world.dimension.OtherworldDimension;
-import net.minecraft.network.chat.Component;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.server.level.ServerPlayer;
-import net.minecraft.sounds.SoundEvents;
-import net.minecraft.sounds.SoundSource;
-import net.minecraft.util.Mth;
-import net.minecraft.world.effect.MobEffectInstance;
-import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.EquipmentSlot;
-import net.minecraft.world.entity.LivingEntity;
-import net.minecraft.world.entity.player.Inventory;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.phys.Vec3;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.event.TickEvent;
 import net.minecraftforge.event.entity.player.PlayerEvent;
@@ -41,12 +32,14 @@ import java.util.*;
 
 @Mod.EventBusSubscriber(modid = Reference.MOD_ID, bus = Mod.EventBusSubscriber.Bus.FORGE)
 public class TickHandler {
+	private static final Map<String, IActionModeHandler> ACTION_MODE_HANDLERS = new HashMap<>();
+	private static final List<IStatusEffectHandler> STATUS_EFFECT_HANDLERS = new ArrayList<>();
 
     private static final int REGEN_INTERVAL = 20;
     private static final int SYNC_INTERVAL = 10;
     private static final double MEDITATION_BONUS_PER_LEVEL = 0.025;
     private static final double ACTIVE_CHARGE_MULTIPLIER = 1.5;
-	private static int saiyanZenkaiSeconds = 0, masterySeconds = 0;
+	private static int masterySeconds = 0;
 
     private static final Map<UUID, Integer> playerTickCounters = new HashMap<>();
 
@@ -61,8 +54,6 @@ public class TickHandler {
 
         StatsProvider.get(StatsCapability.INSTANCE, serverPlayer).ifPresent(data -> {
             if (!data.getStatus().hasCreatedCharacter()) return;
-
-			handleBioDrainTick(serverPlayer, data);
 
 			if (serverPlayer.hasEffect(MainEffects.STUN.get())) {
 				data.getStatus().setChargingKi(false);
@@ -129,26 +120,6 @@ public class TickHandler {
 				data.getStatus().setAuraActive(false);
 			}
 
-			if (data.getSkills().isSkillActive("fly") && !serverPlayer.isCreative() && !serverPlayer.isSpectator()) {
-				if (serverPlayer.horizontalCollision) {
-					double dx = serverPlayer.getX() - serverPlayer.xOld;
-					double dz = serverPlayer.getZ() - serverPlayer.zOld;
-					double speed = Math.sqrt(dx * dx + dz * dz);
-					double minImpactSpeed = 0.35D;
-
-					if (speed > minImpactSpeed) {
-						float maxHealth = serverPlayer.getMaxHealth();
-						double maxImpactSpeedRef = 1.5D;
-						double factor = (speed - minImpactSpeed) / (maxImpactSpeedRef - minImpactSpeed);
-						factor = Mth.clamp(factor, 0.0, 1.0);
-						float finalPct = (float) Mth.lerp(factor, 0.05f, 0.35f);
-						float damage = maxHealth * finalPct;
-						serverPlayer.hurt(serverPlayer.damageSources().flyIntoWall(), damage);
-						serverPlayer.level().playSound(null, serverPlayer.getX(), serverPlayer.getY(), serverPlayer.getZ(), SoundEvents.PLAYER_HURT, SoundSource.PLAYERS, 1.0F, (float) (0.5F + (factor * 0.5F)));
-					}
-				}
-			}
-
 			if (tickCounter % 5 == 0) {
 				boolean hasYajirobe = serverPlayer.getInventory().hasAnyOf(Set.of(MainItems.KATANA_YAJIROBE.get()));
 				boolean holdingYajirobe = serverPlayer.getMainHandItem().getItem() == MainItems.KATANA_YAJIROBE.get() || serverPlayer.getOffhandItem().getItem() == MainItems.KATANA_YAJIROBE.get();
@@ -181,10 +152,12 @@ public class TickHandler {
 
 			}
 
+			for (IStatusEffectHandler handler : STATUS_EFFECT_HANDLERS) {
+				handler.onPlayerTick(serverPlayer, data);
+			}
+
 			if (tickCounter % 20 == 0) {
 				handleActionCharge(serverPlayer, data);
-				handleKaiokenEffects(serverPlayer, data);
-				handleFlightKiDrain(serverPlayer, data);
 				GravityLogic.tick(serverPlayer);
 				if (ConfigManager.getServerConfig().getWorldGen().isOtherworldActive()) {
 					if (!data.getStatus().isAlive() && !serverPlayer.serverLevel().dimension().equals(OtherworldDimension.OTHERWORLD_KEY)) {
@@ -199,21 +172,17 @@ public class TickHandler {
 					data.getCharacter().setActiveForm("androidforms", "androidbase");
 					serverPlayer.refreshDimensions();
 				}
-			}
 
-			if (ConfigManager.getServerConfig().getRacialSkills().isEnableRacialSkills() && ConfigManager.getServerConfig().getRacialSkills().isSaiyanRacialSkill()) {
-				if (tickCounter % 20 == 0 && data.getCharacter().getRaceName().equals("saiyan")) {
-					handleSaiyanPassive(serverPlayer, data);
+				for (IStatusEffectHandler handler : STATUS_EFFECT_HANDLERS) {
+					handler.onPlayerSecond(serverPlayer, data);
 				}
 			}
 
-			fusionTickHandling(serverPlayer, data);
-			handleStatusEffects(serverPlayer, data);
+			if (shouldSync) {
+				NetworkHandler.sendToTrackingEntityAndSelf(new StatsSyncS2C(serverPlayer), serverPlayer);
+			}
+		});
 
-            if (shouldSync) {
-                NetworkHandler.sendToTrackingEntityAndSelf(new StatsSyncS2C(serverPlayer), serverPlayer);
-            }
-        });
     }
 
     @SubscribeEvent
@@ -373,50 +342,9 @@ public class TickHandler {
 		int increment = 0;
 		boolean execute = false;
 
-		switch (mode) {
-			case KAIOKEN -> {
-				if (TransformationsHelper.canStackKaioken(data)) {
-					int skillLvl = data.getSkills().getSkillLevel("kaioken");
-					int currentPhase = data.getStatus().getActiveKaiokenPhase();
-					int maxPhase = TransformationsHelper.getMaxKaiokenPhase(skillLvl);
-					if (currentPhase < maxPhase) {
-						increment = 25;
-					}
-				} else {
-					data.getStatus().setActionCharging(false);
-					return;
-				}
-			}
-			case FUSION -> {
-				if (data.getSkills().hasSkill("fusion") && !data.getCooldowns().hasCooldown(Cooldowns.COMBAT) && !data.getCooldowns().hasCooldown(Cooldowns.FUSION_CD)) {
-					increment = 10;
-				}
-			}
-			case RACIAL -> {
-				String race = data.getCharacter().getRaceName();
-				if ("bioandroid".equals(race)) {
-					RacialSkillLogic.attemptRacialAction(player);
-					data.getStatus().setActionCharging(false);
-					return;
-				}
-				increment = 25;
-			}
-			case FORM -> {
-				FormConfig.FormData nextForm = TransformationsHelper.getNextAvailableForm(data);
-				if (nextForm != null) {
-					String group = data.getCharacter().hasActiveForm() ? data.getCharacter().getActiveFormGroup() : data.getCharacter().getSelectedFormGroup();
-
-					String type = ConfigManager.getFormGroup(data.getCharacter().getRaceName(), group).getFormType();
-					int skillLvl = switch (type) {
-						case "super" -> data.getSkills().getSkillLevel("superform");
-						case "god" -> data.getSkills().getSkillLevel("godform");
-						case "legendary" -> data.getSkills().getSkillLevel("legendaryforms");
-						case "android" -> data.getSkills().getSkillLevel("androidforms");
-						default -> 1;
-					};
-					increment = 5 * Math.max(1, skillLvl);
-				}
-			}
+		IActionModeHandler handler = ACTION_MODE_HANDLERS.get(mode.name());
+		if (handler != null) {
+			increment += handler.handleActionCharge(player, data);
 		}
 
 		if (increment > 0) {
@@ -447,322 +375,40 @@ public class TickHandler {
 	}
 
 	private static boolean performAction(ServerPlayer player, StatsData data, ActionMode mode) {
-		switch (mode) {
-			case KAIOKEN -> {
-				int currentPhase = data.getStatus().getActiveKaiokenPhase();
-				int skillLvl = data.getSkills().getSkillLevel("kaioken");
-				int maxPhase = TransformationsHelper.getMaxKaiokenPhase(skillLvl);
-
-				if (currentPhase < maxPhase) {
-					data.getStatus().setActiveKaiokenPhase(currentPhase + 1);
-					String name = TransformationsHelper.getKaiokenName(currentPhase + 1);
-					player.displayClientMessage(Component.translatable("message.dragonminez.kaioken.activate", name), true);
-				}
-
-				if (!data.getSkills().isSkillActive("kaioken")) data.getSkills().setSkillActive("kaioken", true);
-				NetworkHandler.sendToTrackingEntityAndSelf(new StatsSyncS2C(player), player);
-				return true;
-			}
-			case RACIAL -> {
-				RacialSkillLogic.attemptRacialAction(player);
-				return true;
-			}
-			case FUSION -> {
-				return attemptFusion(player, data);
-			}
-			case FORM -> {
-				attemptTransform(player, data);
-				return true;
-			}
+		IActionModeHandler handler = ACTION_MODE_HANDLERS.get(mode.name());
+		if (handler != null) {
+			return handler.performAction(player, data);
 		}
 		return false;
 	}
 
-	private static void handleKaiokenEffects(ServerPlayer player, StatsData data) {
-		if (!data.getSkills().isSkillActive("kaioken")) return;
-		if (player.isCreative() || player.isSpectator()) return;
-
-		float drain = Math.max(2, TransformationsHelper.getKaiokenHealthDrain(data));
-
-		if (player.getHealth() - drain <= 1.0f) {
-			data.getSkills().setSkillActive("kaioken", false);
-			data.getStatus().setActiveKaiokenPhase(0);
-			data.getResources().setPowerRelease(0);
-			data.getResources().setActionCharge(0);
-		} else {
-			player.setHealth(player.getHealth() - drain);
-		}
+	public static void registerActionModeHandlers() {
+		ACTION_MODE_HANDLERS.put(ActionMode.FORM.name(), new FormModeHandler());
+		ACTION_MODE_HANDLERS.put(ActionMode.FUSION.name(), new FusionModeHandler());
+		ACTION_MODE_HANDLERS.put(ActionMode.KAIOKEN.name(), new KaiokenModeHandler());
+		ACTION_MODE_HANDLERS.put(ActionMode.RACIAL.name(), new RacialModeHandler());
 	}
 
-
-	private static void attemptTransform(ServerPlayer player, StatsData data) {
-		FormConfig.FormData nextForm = TransformationsHelper.getNextAvailableForm(data);
-		if (nextForm == null) return;
-
-		int cost = (int) (data.getMaxEnergy() * nextForm.getEnergyDrain());
-		if (data.getResources().getCurrentEnergy() >= cost) {
-			data.getResources().removeEnergy(cost);
-
-			String group = data.getCharacter().hasActiveForm() ?
-					data.getCharacter().getActiveFormGroup() :
-					data.getCharacter().getSelectedFormGroup();
-
-			data.getCharacter().setActiveForm(group, nextForm.getName());
-			player.refreshDimensions();
-
-            player.level().playSound(null, player.getX(), player.getY(), player.getZ(), MainSounds.TRANSFORM.get(), SoundSource.PLAYERS, 1.0F, 1.0F);
-		} else {
-			player.displayClientMessage(Component.translatable("message.dragonminez.form.no_ki", cost), true);
-		}
+	public static void registerStatusEffectHandlers() {
+		STATUS_EFFECT_HANDLERS.add(new BioDrainHandler());
+		STATUS_EFFECT_HANDLERS.add(new DashStatusHandler());
+		STATUS_EFFECT_HANDLERS.add(new DoubleDashStatusHandler());
+		STATUS_EFFECT_HANDLERS.add(new FlyStatusHandler());
+		STATUS_EFFECT_HANDLERS.add(new FusionStatusHandler());
+		STATUS_EFFECT_HANDLERS.add(new KaiokenStatusHandler());
+		STATUS_EFFECT_HANDLERS.add(new KiChargeStatusHandler());
+		STATUS_EFFECT_HANDLERS.add(new MajinStatusHandler());
+		STATUS_EFFECT_HANDLERS.add(new MightFruitStatusHandler());
+		STATUS_EFFECT_HANDLERS.add(new SaiyanPassiveHandler());
+		STATUS_EFFECT_HANDLERS.add(new TransformStatusHandler());
 	}
 
-	private static boolean attemptFusion(ServerPlayer player, StatsData data) {
-		List<ServerPlayer> nearby = player.level().getEntitiesOfClass(ServerPlayer.class,
-				player.getBoundingBox().inflate(5.0), p -> p != player);
-
-		for (ServerPlayer target : nearby) {
-			StatsProvider.get(StatsCapability.INSTANCE, target).ifPresent(targetData -> {
-				if (targetData.getStatus().getSelectedAction() == ActionMode.FUSION && targetData.getResources().getActionCharge() >= 50 && targetData.getStatus().isActionCharging()) {
-					if (data.getResources().getActionCharge() >= 100) {
-						if (FusionLogic.executeMetamoru(player, target, data, targetData)) {
-							data.getResources().setActionCharge(0);
-							targetData.getResources().setActionCharge(0);
-
-                            player.level().playSound(null, player.getX(), player.getY(), player.getZ(), MainSounds.FUSION.get(), SoundSource.PLAYERS, 1.0F, 1.0F);
-						}
-					}
-				}
-			});
-			if (data.getStatus().isFused()) return true;
-		}
-		return false;
+	public static void registerActionModeHandler(String actionMode, IActionModeHandler actionModeHandler) {
+		// FIXME: Should we write an error to the log when someone attempts to register a handler for an existing mode?
+		ACTION_MODE_HANDLERS.putIfAbsent(actionMode, actionModeHandler);
 	}
 
-	private static void handleSaiyanPassive(ServerPlayer player, StatsData data) {
-		GeneralServerConfig.RacialSkillsConfig config = ConfigManager.getServerConfig().getRacialSkills();
-
-		if (data.getResources().getRacialSkillCount() >= config.getSaiyanZenkaiAmount()) return;
-		if (data.getCooldowns().hasCooldown(Cooldowns.ZENKAI)) return;
-
-		float maxHealth = player.getMaxHealth();
-		if (player.getHealth() <= maxHealth * 0.15) {
-			saiyanZenkaiSeconds = saiyanZenkaiSeconds + 1;
-		} else {
-			saiyanZenkaiSeconds = 0;
-		}
-
-		if (saiyanZenkaiSeconds >= 8) {
-			player.heal((float) (maxHealth * config.getSaiyanZenkaiHealthRegen()));
-
-			double boostMult = config.getSaiyanZenkaiStatBoost();
-			String[] statsToBoost = config.getSaiyanZenkaiBoosts();
-
-			for (String statKey : statsToBoost) {
-				int currentStat = getStatValue(data, statKey);
-				int bonus = (int) Math.max(1, currentStat * boostMult);
-				data.getBonusStats().addBonus(statKey, "Zenkai_" + (data.getResources().getRacialSkillCount() + 1), "+", bonus);
-			}
-
-			player.displayClientMessage(Component.translatable("message.dragonminez.racial.zenkai.used"), true);
-
-			data.getResources().addRacialSkillCount(1);
-			data.getCooldowns().setCooldown(Cooldowns.ZENKAI, config.getSaiyanZenkaiCooldownSeconds());
-			NetworkHandler.sendToTrackingEntityAndSelf(new StatsSyncS2C(player), player);
-			saiyanZenkaiSeconds = 0;
-		}
-	}
-
-	private static int getStatValue(StatsData data, String statName) {
-		return switch (statName) {
-			case "STR" -> data.getStats().getStrength();
-			case "SKP" -> data.getStats().getStrikePower();
-			case "RES" -> data.getStats().getResistance();
-			case "VIT" -> data.getStats().getVitality();
-			case "PWR" -> data.getStats().getKiPower();
-			case "ENE" -> data.getStats().getEnergy();
-			default -> 0;
-		};
-	}
-
-	private static void handleBioDrainTick(ServerPlayer player, StatsData data) {
-		int targetId = data.getStatus().getDrainingTargetId();
-		if (targetId == -1) return;
-
-		if (data.getCooldowns().getCooldown(Cooldowns.DRAIN_ACTIVE) > 0) {
-			Entity entity = player.level().getEntity(targetId);
-			if (entity instanceof LivingEntity target && target.isAlive() && player.distanceTo(target) < 6.0f) {
-				float targetYRot = target.getYRot();
-
-				double dist = 0.75;
-				double rads = Math.toRadians(targetYRot);
-				double xOffset = -Math.sin(rads) * dist;
-				double zOffset = Math.cos(rads) * dist;
-
-				double finalX = target.getX() - xOffset;
-				double finalZ = target.getZ() - zOffset;
-				double finalY = target.getY();
-
-				player.connection.teleport(finalX, finalY, finalZ, targetYRot, 15.0F);
-
-				player.setYRot(targetYRot);
-				player.setYHeadRot(targetYRot);
-
-				if (player.tickCount % 20 == 0) {
-					GeneralServerConfig.RacialSkillsConfig config = ConfigManager.getServerConfig().getRacialSkills();
-					double totalDrainRatio = config.getBioAndroidDrainRatio();
-
-					float durationSeconds = 6.0f;
-					float totalHealthToDrain = (float) (target.getMaxHealth() * totalDrainRatio);
-					float drainPerSecond = totalHealthToDrain / durationSeconds;
-
-					target.setHealth(Math.max(1, target.getHealth() - drainPerSecond));
-					if (player.getHealth() < player.getMaxHealth()) {
-						player.heal(drainPerSecond);
-					}
-					if (data.getResources().getCurrentEnergy() < data.getMaxEnergy()) {
-						data.getResources().addEnergy((int)(drainPerSecond * 5));
-					}
-					target.playSound(MainSounds.ABSORB1.get());
-					player.playSound(MainSounds.ABSORB1.get());
-
-					if (target.getHealth() <= 1.0f) {
-						target.hurtMarked = true;
-						data.getStatus().setDrainingTargetId(-1);
-						data.getCooldowns().removeCooldown(Cooldowns.DRAIN_ACTIVE);
-						player.removeEffect(MainEffects.STUN.get());
-						target.kill();
-					}
-				}
-
-			} else {
-				data.getStatus().setDrainingTargetId(-1);
-				data.getCooldowns().removeCooldown(Cooldowns.DRAIN_ACTIVE);
-				player.removeEffect(MainEffects.STUN.get());
-			}
-
-		} else {
-			Entity entity = player.level().getEntity(targetId);
-			if (entity instanceof LivingEntity target) {
-				Vec3 look = player.getLookAngle();
-				target.setDeltaMovement(look.scale(1.5).add(0, 0.5, 0));
-				target.hurtMarked = true;
-				player.setDeltaMovement(look.scale(-1.0).add(0, 0.3, 0));
-				player.hurtMarked = true;
-				target.playSound(MainSounds.KNOCKBACK_CHARACTER.get());
-				player.playSound(MainSounds.KNOCKBACK_CHARACTER.get());
-			}
-			data.getStatus().setDrainingTargetId(-1);
-			player.removeEffect(MainEffects.STUN.get());
-		}
-	}
-
-	private static void fusionTickHandling(ServerPlayer serverPlayer, StatsData data) {
-		if (data.getStatus().isFused()) {
-			if (data.getStatus().isFusionLeader()) {
-				int timer = data.getStatus().getFusionTimer();
-				if (timer > 0) {
-					data.getStatus().setFusionTimer(timer - 1);
-					if (timer - 1 <= 0) FusionLogic.endFusion(serverPlayer, data, false);
-				}
-				UUID partnerUUID = data.getStatus().getFusionPartnerUUID();
-				ServerPlayer partner = serverPlayer.getServer().getPlayerList().getPlayer(partnerUUID);
-				if (partner == null || partner.hasDisconnected()) {
-					FusionLogic.endFusion(serverPlayer, data, true);
-				} else if (partner.isDeadOrDying()) {
-					FusionLogic.endFusion(serverPlayer, data, true);
-				} else if (partner.distanceTo(serverPlayer) > 5.0) {
-					partner.teleportTo(serverPlayer.getX(), serverPlayer.getY(), serverPlayer.getZ());
-				}
-			} else {
-				UUID leaderUUID = data.getStatus().getFusionPartnerUUID();
-				ServerPlayer leader = serverPlayer.getServer().getPlayerList().getPlayer(leaderUUID);
-				if (leader == null || leader.hasDisconnected() || leader.isDeadOrDying()) FusionLogic.endFusion(serverPlayer, data, true);
-			}
-		}
-	}
-
-	private static void handleFlightKiDrain(ServerPlayer player, StatsData data) {
-		if (!data.getSkills().isSkillActive("fly")) return;
-		if (player.isCreative() || player.isSpectator()) return;
-
-		int flyLevel = data.getSkills().getSkillLevel("fly");
-		if (flyLevel >= data.getSkills().getMaxSkillLevel("fly")) return;
-		int maxEnergy = data.getMaxEnergy();
-
-		double basePercent = player.isSprinting() ? 0.08 : 0.03;
-		double energyCostPercent = Math.max(0.002, basePercent - (flyLevel * 0.005));
-		int energyCost = (int) Math.ceil(maxEnergy * energyCostPercent);
-
-		int currentEnergy = data.getResources().getCurrentEnergy();
-
-		if (currentEnergy >= energyCost) {
-			data.getResources().removeEnergy(energyCost);
-		} else {
-			data.getSkills().setSkillActive("fly", false);
-			if (!player.isCreative() && !player.isSpectator()) {
-				player.getAbilities().mayfly = false;
-				player.getAbilities().flying = false;
-				player.onUpdateAbilities();
-			}
-			NetworkHandler.sendToTrackingEntityAndSelf(new StatsSyncS2C(player), player);
-		}
-	}
-
-	private static void handleStatusEffects(ServerPlayer player, StatsData data) {
-		if (data.getStatus().isChargingKi()) {
-			if (!player.hasEffect(MainEffects.KICHARGE.get())) {
-				player.addEffect(new MobEffectInstance(MainEffects.KICHARGE.get(), -1, 0, false, false, true));
-			}
-		} else {
-			player.removeEffect(MainEffects.KICHARGE.get());
-		}
-
-		if (data.getStatus().isActionCharging()) {
-			if (!player.hasEffect(MainEffects.TRANSFORM.get())) {
-				player.addEffect(new MobEffectInstance(MainEffects.TRANSFORM.get(), -1, 0, false, false, true));
-			}
-		} else {
-			player.removeEffect(MainEffects.TRANSFORM.get());
-		}
-
-		if (data.getSkills().isSkillActive("kaioken")) {
-			if (!player.hasEffect(MainEffects.KAIOKEN.get())) {
-				player.addEffect(new MobEffectInstance(MainEffects.KAIOKEN.get(), -1, 0, false, false, true));
-			}
-		} else {
-			player.removeEffect(MainEffects.KAIOKEN.get());
-		}
-
-		if (data.getSkills().isSkillActive("fly")) {
-			if (!player.hasEffect(MainEffects.FLY.get())) {
-				player.addEffect(new MobEffectInstance(MainEffects.FLY.get(), -1, 0, false, false, true));
-			}
-		} else {
-			player.removeEffect(MainEffects.FLY.get());
-		}
-
-		if (data.getEffects().hasEffect("majin")) {
-			if (!player.hasEffect(MainEffects.MAJIN.get())) {
-				player.addEffect(new MobEffectInstance(MainEffects.MAJIN.get(), -1, 0, false, false, true));
-			}
-		} else {
-			player.removeEffect(MainEffects.MAJIN.get());
-		}
-
-		if (data.getEffects().hasEffect("mightfruit")) {
-			if (!player.hasEffect(MainEffects.MIGHTFRUIT.get())) {
-				player.addEffect(new MobEffectInstance(MainEffects.MIGHTFRUIT.get(), -1, 0, false, false, true));
-			}
-		} else {
-			player.removeEffect(MainEffects.MIGHTFRUIT.get());
-		}
-		
-		if (!data.getCooldowns().hasCooldown(Cooldowns.DASH_CD)) {
-			player.removeEffect(MainEffects.DASH_CD.get());
-		}
-		
-		if (!data.getCooldowns().hasCooldown(Cooldowns.DOUBLEDASH_CD)) {
-			player.removeEffect(MainEffects.DOUBLEDASH_CD.get());
-		}
+	public static void registerStatusEffectHandler(IStatusEffectHandler statusEffectHandler) {
+		STATUS_EFFECT_HANDLERS.add(statusEffectHandler);
 	}
 }
