+++ src/main/java/com/dragonminez/server/events/players/statuseffect/DashStatusHandler.java
@@ -1,0 +1,26 @@
+package com.dragonminez.server.events.players.statuseffect;
+
+import com.dragonminez.common.init.MainEffects;
+import com.dragonminez.common.stats.Cooldowns;
+import com.dragonminez.common.stats.StatsData;
+import com.dragonminez.server.events.players.IStatusEffectHandler;
+import net.minecraft.server.level.ServerPlayer;
+
+public class DashStatusHandler implements IStatusEffectHandler {
+    @Override
+    public void handleStatusEffects(ServerPlayer player, StatsData data) {
+        if (!data.getCooldowns().hasCooldown(Cooldowns.DASH_CD)) {
+            player.removeEffect(MainEffects.DASH_CD.get());
+        }
+    }
+
+    @Override
+    public void onPlayerTick(ServerPlayer serverPlayer, StatsData data) {
+
+    }
+
+    @Override
+    public void onPlayerSecond(ServerPlayer serverPlayer, StatsData data) {
+
+    }
+}
